* Fixed some clang compiler warnings. * Fixed a dodgey self-comparison in operator<(const Key &other). * Added some const-correctness here and there. * Removed an unused maxRetries variable from sim-unlock-dialog. * Replaced #ifndef's with #pragma once.

+++ src/indicator/nmofono/wwan/modem.cpp
@@ -304,7 +304,7 @@ public Q_SLOTS:
         {
             setOperatorName(m_networkRegistration->name());
             setStatus(str2status(m_networkRegistration->status()));
-            setStrength(m_networkRegistration->strength());
+            setStrength((int8_t)m_networkRegistration->strength());
         }
         else
         {
